+++ server/game/ai/task/actor.go
@@ -1,7 +1,7 @@
 package task
 
 type Actor interface {
-	CreateActor(id int64)
+	CreateActor(id int)
 	FindPath() *[]int64
 	Move(nodeID int64)
 }
